move PagerControl
